+++ network/p2p/p2pconf/errors.go
@@ -4,7 +4,7 @@ import (
 	"errors"
 	"fmt"
 
-	"github.com/onflow/flow-go/network/p2p"
+	"github.com/onflow/flow-go/network/p2p/message"
 )
 
 // InvalidLimitConfigError indicates the validation limit is < 0.
@@ -21,7 +21,7 @@ func (e InvalidLimitConfigError) Unwrap() error {
 }
 
 // NewInvalidLimitConfigErr returns a new ErrValidationLimit.
-func NewInvalidLimitConfigErr(controlMsg p2p.ControlMessageType, err error) InvalidLimitConfigError {
+func NewInvalidLimitConfigErr(controlMsg p2pmsg.ControlMessageType, err error) InvalidLimitConfigError {
 	return InvalidLimitConfigError{fmt.Errorf("invalid rpc control message %s validation limit configuration: %w", controlMsg, err)}
 }
